+++ packages/types/src/module.ts
@@ -8,6 +8,8 @@ import type { RpcRequest, RpcResponse } from './rpc';
 import type { NetworkContractToken } from './token';
 import type { GetTransactionHistory, TransactionHistoryResponse } from './transaction-history';
 import type { ApprovalController } from './rpc';
+import type { HyperSDKClient } from 'hypersdk-client';
+import { type Rpc, type SolanaRpcApiDevnet, type SolanaRpcApiMainnet } from '@solana/rpc';
 
 export type AppInfo = {
   name: AppName;
@@ -23,7 +25,15 @@ export type ConstructorParams = {
 export type NetworkFeeParam = Network & { caipId?: string };
 
 export interface Module {
-  getProvider: (network: Network) => Promise<JsonRpcBatchInternal | BitcoinProvider | Avalanche.JsonRpcProvider>;
+  getProvider: (
+    network: Network,
+  ) => Promise<
+    | JsonRpcBatchInternal
+    | BitcoinProvider
+    | Avalanche.JsonRpcProvider
+    | HyperSDKClient
+    | Rpc<SolanaRpcApiMainnet | SolanaRpcApiDevnet>
+  >;
   getManifest: () => Manifest | undefined;
   getBalances: (params: GetBalancesParams) => Promise<GetBalancesResponse>;
   getTransactionHistory: (params: GetTransactionHistory) => Promise<TransactionHistoryResponse>;