+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @avalabs/vm-module-types
 
+## 0.12.1
+
+### Patch Changes
+
+- 20e096d: chore: update SDKs
+
+## 0.12.0
+
+### Minor Changes
+
+- b504f6f: new networktype
+
 ## 0.11.14
 
 ## 0.11.13